refactor(evals): extract score() to evals/scoring.py for import decoupling
score() is pure logic over captured CaseResult data; it has no runtime
dependency on anthropic or mcp. Keeping it in evals/harness.py forced
tests/test_evals_score.py to transitively import anthropic, breaking
CI where the evals dependency group isn't installed.

New module boundary:
- evals/scoring.py — pure-Python scoring layer, no runtime deps.
- evals/harness.py — runtime harness that orchestrates live LLM + MCP
  calls and uses the scoring layer.

Function body is moved verbatim; no behavior change.
